Module magnetic::mpsc

Multiple-producer single-consumer queue

The MPSC queue allows for pushing from one thread and popping from another. The producer end of the queue may be accessed by multiple threads while the consumer end may only be accessed by a single thread. In other words, the MPSCProducer is Send and Sync while the MPSCConsumer is Send and !Sync.

Structs

MPSCConsumer
Consumer end of the queue. Implements the trait Consumer<T>.

MPSCProducer
Producer end of the queue. Implements the trait Producer<T>.

Functions

mpsc_queue
Creates a new MPSC queue