Trait magnetic::Producer

pub trait Producer<T> {
    fn push(&self, value: T) -> Result<(), PushError<T>>;
    fn try_push(&self, value: T) -> Result<(), TryPushError<T>>;
}

The consumer end of the queue allows for sending data. Producer<T> is always Send, but is only Sync for multi-producer (MPSC, MPMC) queues.

Required Methods

fn push(&self, value: T) -> Result<(), PushError<T>>

Add value to front of the queue. This method will block if the queue is currently full.

fn try_push(&self, value: T) -> Result<(), TryPushError<T>>

Attempt to add a value to the front of the queue. If the value was added successfully, None will be returned. If unsuccessful, value will be returned. An unsuccessful push indicates that the queue was full.